# Conflicts:
#	_db/changes-log.sql

+++ _db/changes-log.sql
@@ -113,7 +113,7 @@ CHANGE `issue_sp` `issue_sp` varchar(255) COLLATE 'utf8_general_ci' NOT NULL COM
 
 -- 2017-12-01 13:01:00
 
-ALTER TABLE `lpm_issues` CHANGE `hours` `hours` FLOAT NOT NULL;
+ALTER TABLE `lpm_issues` CHANGE `hours` `hours` DECIMAL(10,1) NOT NULL;
 
 -- 2017-12-01 14:12:00
